Moved all character tracking commands from Admin cog to Tracking cog - Standardized /im, /claim and /addacc - Merged settins cog into Admin cog
